simplify linear systems driver
